리뷰 작성 및 평점 계산 추가

+++ app/src/main/java/com/tistory/puzzleleaf/rankofalcohol/model/FbUser.java
@@ -5,7 +5,7 @@ package com.tistory.puzzleleaf.rankofalcohol.model;
  */
 
 public class FbUser {
-    private float hMany;
+    private double hMany;
     private String gender;
 
     public FbUser(){}
@@ -15,7 +15,7 @@ public class FbUser {
         this.gender = gender;
     }
 
-    public float gethMany() {
+    public double gethMany() {
         return hMany;
     }
